+++ src/ext.rs
@@ -5,3 +5,7 @@ mod num_bigint;
 #[cfg(feature = "crypto-bigint")]
 #[cfg_attr(docsrs, doc(cfg(feature = "crypto-bigint")))]
 mod crypto_bigint;
+
+#[cfg(feature = "rug")]
+#[cfg_attr(docsrs, doc(cfg(feature = "rug")))]
+mod rug;