+++ src/tests.rs
@@ -4,5 +4,5 @@ mod rundgren;
 mod simple_value;
 mod value;
 
-#[cfg(all(feature = "num-bigint", feature = "crypto-bigint"))]
+#[cfg(all(feature = "num-bigint", feature = "crypto-bigint", feature = "rug"))]
 mod bigint_interop;